Terminar caso de uso

+++ backend/database/migrations/2020_10_20_165442_eliminar_id_respuesta_ofrecida_opcion_evaluacion.php
@@ -13,11 +13,13 @@ class EliminarIdRespuestaOfrecidaOpcionEvaluacion extends Migration
      */
     public function up()
     {
-        Schema::table('opcion_evaluacion', function (Blueprint $table) {
-            $table->dropForeign(['idRespuestaOfrecida']);
-            $table->dropColumn('idRespuestaOfrecida');
-            $table->dropColumn('idRespuestaOfrecida');
-        });
+        if (Schema::hasColumn('opcion_evaluacion', 'idRespuestaOfrecida'))
+        {
+            Schema::table('opcion_evaluacion', function (Blueprint $table) {
+                $table->dropForeign(['idRespuestaOfrecida']);
+                $table->dropColumn('idRespuestaOfrecida');
+            });
+        }
     }
 
     /**
